pyproject.toml
[project]
name = "project1-skypro"
version = "0.1.0"
description = ""
authors = [
    {name = "Margarita",email = "mbursheva@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.17.1"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^7.0.0"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119


